+++ baybe/exceptions.py
@@ -45,4 +45,8 @@ class NumericalUnderflowError(Exception):
 
 
 class OptionalImportError(Exception):
-    """Attempted to import an optional but uninstalled dependency."""
+    """An attempt was made to import an optional but uninstalled dependency."""
+
+
+class DeprecationError(Exception):
+    """Can be used to signal deprecations to the user, interrupting code exception."""