+++ CHANGELOG.md
@@ -19,6 +19,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - `botorch_function_wrapper` from `baybe.utils` namespace
 
+### Deprecations
+- Passing a `numerical_measurements_must_be_within_tolerance` to the `Campaign` 
+  constructor is no longer supported. Instead, `Campaign.add_measurements` now
+  takes an additional parameter to control the behavior.
+
 ## [0.7.2] - 2024-01-24
 ### Added
 - Target enums